backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
fastapi==0.109.0
uvicorn==0.27.0

# OCR and Image Processing
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# API and Validation
pydantic==2.5.3
python-multipart==0.0.6
requests==2.31.0

# ML and Data Processing
scikit-learn==1.4.0
pandas==2.2.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
PyYAML==6.0.1
